Extract pivot TP alignment logic to TakeProfitCalculator service
